refactor: cryptoUtils
