+++ lib/swaps/Swaps.ts
@@ -10,8 +10,9 @@ import { PacketType } from '../p2p/packets';
 import * as packets from '../p2p/packets/types';
 import Peer from '../p2p/Peer';
 import Pool from '../p2p/Pool';
+import { generatePreimageAndHash } from '../utils/cryptoUtils';
 import { UnitConverter } from '../utils/UnitConverter';
-import { generatePreimageAndHash, setTimeoutPromise } from '../utils/utils';
+import { setTimeoutPromise } from '../utils/utils';
 import { MAX_PAYMENT_TIME } from './consts';
 import errors, { errorCodes } from './errors';
 import SwapClient, { PaymentState } from './SwapClient';
